refactor: apis 구조 변경
