+++ apis/auth.ts
@@ -1,4 +1,4 @@
-import { client, type MutationRes } from '../client'
+import { client, type MutationRes } from '.'
 
 export interface PostLoginProps {
   login_id: string
