--- apis/user/getMe.ts
@@ -1,3 +1,0 @@
-import { userClient } from '../client'
-
-export const getMe = userClient.get<null>('/me')
