--- apis/user/postCode.ts
@@ -1,8 +1,0 @@
-import { userClient } from '../client'
-
-export interface PostCodeProps {
-  email: string
-}
-
-export const postCode = async (props: PostCodeProps) =>
-  await userClient.post('/verification-code', props)
