--- apis/user/postPassword.ts
@@ -1,8 +1,0 @@
-import { type MutationRes, userClient } from '../client'
-
-interface PostPasswordProps {
-  password: string
-}
-
-export const postPassword = async (props: PostPasswordProps) =>
-  await userClient.post<MutationRes, PostPasswordProps>('/password', props)
