+++ hooks/auth/mutations/useEmailCode.ts
@@ -1,6 +1,6 @@
 import { useMutation } from '@tanstack/react-query'
 
-import { postCode, type PostCodeProps } from '@/apis/user/postCode'
+import { postCode, type PostCodeProps } from '@/apis/user'
 
 export const useEmailCode = () =>
   useMutation({
