+++ hooks/auth/mutations/useLogin.ts
@@ -1,6 +1,6 @@
 import { useMutation } from '@tanstack/react-query'
 
-import { postLogin, type PostLoginProps } from '@/apis/auth/postLogin'
+import { postLogin, type PostLoginProps } from '@/apis/auth'
 
 export const useLogin = () =>
   useMutation({
